Crate cirrus_primitives

Primitives for executor pallet.

Traits

SecondaryApi

API necessary for secondary node.

Signer

Extracts the signer from an unchecked extrinsic.

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Address

The address format for describing accounts.

Balance

Balance of an account.

BlockNumber

An index to a block.

Hash

A hash of some data used by the chain.

Index

Index of a transaction in the chain.

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.